# python-sdk/openapi.yaml
openapi: 3.0.0
info:
  contact:
    email: support@katanemo.com
    name: Katanemo Labs Inc.
    url: https://api.katanemo.com
  description: "With Katanemo developers can add support for users, enterprise SSO,\
    \ machine keys and fine-grained authorization in minutes."
  termsOfService: https://www.katanemo.com/terms-of-service
  title: "Katanemo - identity, and fine-grained authorization for modern { API-first\
    \ } software companies."
  version: 1.0.0
servers:
- url: https://api.katanemo.com
...
paths:
...
  /org/{accountId}/user:
    get:
      description: Returns a list of all users belonging to provided organization
        ID
      operationId: GetUsersForAccount
      parameters:
      - explode: false
        in: path
        name: accountId
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/UserResponse'
                type: array
          description: users belonging to provided account ID
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadRequestException'
          description: Bad Request Exception
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnauthorizedException'
          description: Unauthorized Exception
        "409":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConflictException'
          description: Conflict Exception
        "429":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TooManyRequestsException'
          description: Too Many Requests Exception
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalServerErrorException'
          description: Internal Server Error
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: unexpected error
      summary: List Users
      tags:
      - organization
      - identity
    post:
      description: "Creates a new User and triggers an email verification workflow,\
        \ followed by set-password"
      operationId: createUserForAccount
      parameters:
      - explode: false
        in: path
        name: accountId
        required: true
        schema:
          type: string
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRequest'
        description: User information
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
          description: User
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadRequestException'
          description: Bad Request Exception
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnauthorizedException'
          description: Unauthorized Exception
        "409":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConflictException'
          description: Conflict Exception
        "429":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TooManyRequestsException'
          description: Too Many Requests Exception
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalServerErrorException'
          description: Internal Server Error
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: unexpected error
      summary: Invite User
      tags:
      - organization
      - identity
  /org/{accountId}/user/{userId}:
    get:
      description: Get a specific user for a particular organization
      operationId: GetUser
      parameters:
      - explode: false
        in: path
        name: accountId
        required: true
        schema:
          type: string
        style: simple
      - explode: false
        in: path
        name: userId
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
          description: user belonging to provided account ID
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadRequestException'
          description: Bad Request Exception
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnauthorizedException'
          description: Unauthorized Exception
        "409":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConflictException'
          description: Conflict Exception
        "429":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TooManyRequestsException'
          description: Too Many Requests Exception
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalServerErrorException'
          description: Internal Server Error
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: unexpected error
      summary: Get User
      tags:
      - organization
      - identity
    put:
      description: Updates a User account
      operationId: updateUser
      parameters:
      - explode: false
        in: path
        name: userId
        required: true
        schema:
          type: string
        style: simple
      - explode: false
        in: path
        name: accountId
        required: true
        schema:
          type: string
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRequest'
        description: User information
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
          description: User
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadRequestException'
          description: Bad Request Exception
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnauthorizedException'
          description: Unauthorized Exception
        "409":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConflictException'
          description: Conflict Exception
        "429":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TooManyRequestsException'
          description: Too Many Requests Exception
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalServerErrorException'
          description: Internal Server Error
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: unexpected error
      summary: Update user
      tags:
      - organization
      - identity
...
  /org/{accountId}/keys:
    get:
      description: List all client keys for an organization accessing a developers
        service
      operationId: getClientKeyList
      parameters:
      - explode: false
        in: path
        name: accountId
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/ClientKeyObject'
                type: array
          description: List of all users successful.
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadRequestException'
          description: Bad Request Exception
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnauthorizedException'
          description: Unauthorized Exception
        "409":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConflictException'
          description: Conflict Exception
        "429":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TooManyRequestsException'
          description: Too Many Requests Exception
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalServerErrorException'
          description: Internal Server Error
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: unexpected error
      summary: List API Keys
      tags:
      - organization
      - identity
      - access-control
    post:
      description: Creates a new client key for accessing a developers APIs
      operationId: createClientKey
      parameters:
      - explode: false
        in: path
        name: accountId
        required: true
        schema:
          type: string
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ClientKeyRequest'
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClientKeyResponse'
          description: Successful org keys.
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadRequestException'
          description: Bad Request Exception
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnauthorizedException'
          description: Unauthorized Exception
        "409":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConflictException'
          description: Conflict Exception
        "429":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TooManyRequestsException'
          description: Too Many Requests Exception
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalServerErrorException'
          description: Internal Server Error
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: unexpected error
      summary: Create API Key
      tags:
      - organization
      - identity
      - access-control
  /org/{accountId}/key/{keyId}:
    delete:
      description: Delete a particular API Key for an organization.
      operationId: deleteClientKey
      parameters:
      - explode: false
        in: path
        name: accountId
        required: true
        schema:
          type: string
        style: simple
      - explode: false
        in: path
        name: keyId
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                type: string
          description: Successful deletion of client key.
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadRequestException'
          description: Bad Request Exception
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnauthorizedException'
          description: Unauthorized Exception
        "409":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConflictException'
          description: Conflict Exception
        "429":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TooManyRequestsException'
          description: Too Many Requests Exception
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalServerErrorException'
          description: Internal Server Error
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: unexpected error
      summary: Delete API Key
      tags:
      - organization
      - identity
      - access-control
    get:
      description: Get details of a particular API key for an organization.
      operationId: getClientKey
      parameters:
      - explode: false
        in: path
        name: accountId
        required: true
        schema:
          type: string
        style: simple
      - explode: false
        in: path
        name: keyId
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClientKeyObject'
          description: Getting client key successful.
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadRequestException'
          description: Bad Request Exception
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnauthorizedException'
          description: Unauthorized Exception
        "409":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConflictException'
          description: Conflict Exception
        "429":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TooManyRequestsException'
          description: Too Many Requests Exception
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalServerErrorException'
          description: Internal Server Error
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: unexpected error
      summary: Get API Key
      tags:
      - organization
      - identity
      - access-control
...
components:
  schemas:
...
    UserRequest:
      example:
        accountId: accountId
        roles:
        - roleId1
        - roleId2
        - roleId3
        userId: userId
        tags:
          key:
          - tags
          - tags
      properties:
        accountId:
          description: Tenant name
          title: accountId
          type: string
        userId:
          description: email address of the user
          title: userId
          type: string
        tags:
          additionalProperties:
            items:
              type: string
            type: array
          title: tags
          type: object
        roles:
          example:
          - roleId1
          - roleId2
          - roleId3
          items:
            type: string
          title: roles
          type: array
      required:
      - accountId
      title: UserRequest
    UserResponse:
      example:
        accountId: accountId
        roles:
        - roleId1
        - roleId2
        - roleId3
        isAdmin: true
        serviceId: serviceId
        isActive: true
        userId: userId
        token: token
        tags:
          key:
          - tags
          - tags
      properties:
        accountId:
          description: Organization Identifier for a SaaS customer
          title: accountId
          type: string
        serviceId:
          description: Service which this account subscribed to
          title: serviceId
          type: string
        userId:
          description: Email
          title: userId
          type: string
        isAdmin:
          description: is the user an admin?
          title: isAdmin
          type: boolean
        isActive:
          description: Is the user active yet?
          title: isActive
          type: boolean
        token:
          description: User token
          title: token
          type: string
        tags:
          additionalProperties:
            items:
              type: string
            type: array
          title: tags
          type: object
        roles:
          example:
          - roleId1
          - roleId2
          - roleId3
          items:
            type: string
          title: roles
          type: array
      required:
      - accountId
      title: UserResponse
...
    ClientKeyRequest:
      example:
        clientName: clientName
        defaultRoleId: defaultRoleId
      properties:
        defaultRoleId:
          title: defaultRoleId
          type: string
        clientName:
          title: clientName
          type: string
      required:
      - clientName
      - defaultRoleId
      title: ClientKeyRequest
    ClientKeyResponse:
      example:
        clientId: clientId
        clientSecret: clientSecret
      properties:
        clientId:
          title: clientId
          type: string
        clientSecret:
          title: clientSecret
          type: string
      title: ClientKeyResponse
    ClientKeyObject:
      example:
        accountId: accountId
        clientName: clientName
        roles:
        - roleId1
        - roleId2
        - roleId3
        clientKeyId: clientKeyId
        serviceId: serviceId
        isActive: true
      properties:
        accountId:
          title: accountId
          type: string
        roles:
          example:
          - roleId1
          - roleId2
          - roleId3
          items:
            type: string
          title: roles
          type: array
        clientName:
          title: clientName
          type: string
        clientKeyId:
          title: clientKeyId
          type: string
        serviceId:
          title: serviceId
          type: string
        isActive:
          title: isActive
          type: boolean
      title: ClientKeyObject
...
    Error:
      example:
        code: 0
        message: message
      properties:
        code:
          description: Error code
          format: int32
          title: code
          type: integer
        message:
          description: Error message
          title: message
          type: string
      required:
      - code
      - message
      title: Error
    BadRequestException:
      properties:
        message:
          title: message
          type: string
        errorCode:
          title: errorCode
          type: string
      title: BadRequestException
      type: object
    UnauthorizedException:
      properties:
        message:
          title: message
          type: string
        errorCode:
          title: errorCode
          type: string
      title: UnauthorizedException
      type: object
...
    ConflictException:
      properties:
        message:
          title: message
          type: string
        errorCode:
          title: errorCode
          type: string
      title: ConflictException
      type: object
    TooManyRequestsException:
      properties:
        message:
          title: message
          type: string
        errorCode:
          title: errorCode
          type: string
      title: TooManyRequestsException
      type: object
    InternalServerErrorException:
      properties:
        message:
          title: message
          type: string
        errorCode:
          title: errorCode
          type: string
      title: InternalServerErrorException
      type: object
...